refactor: move model related to release in dedicated folder
