+++ api/bin/www
@@ -30,7 +30,7 @@ async function exitOnSignal(signal) {
     await checkUrlQueue.close();
     await releaseJob.queue.close();
     process.exit(0);
-  } catch(err) {
+  } catch (err) {
     logger.error(err);
     process.exit(1);
   }
