+++ api/lib/domain/models/release/CourseForRelease.js
