+++ api/tests/tooling/domain-builder/factory/build-skill-for-release.js
@@ -1,4 +1,4 @@
-const SkillForRelease = require('../../../../lib/domain/models/SkillForRelease');
+const SkillForRelease = require('../../../../lib/domain/models/release/SkillForRelease');
 
 const buildSkillForRelease = function({
   id = 'recTIddrkopID28Ep',
